[Move] configService from `src/code/common` to `src/code/platform`
