+++ src/code/electron/nota.ts
@@ -4,7 +4,7 @@ import { ErrorHandler } from "src/base/common/error";
 import { Event } from "src/base/common/event";
 import { ILogService } from "src/base/common/logger";
 import { getUUID, UUID } from "src/base/node/uuid";
-import { IGlobalConfigService, IUserConfigService } from "src/code/common/service/configService/configService";
+import { IGlobalConfigService, IUserConfigService } from "src/code/platform/configuration/electron/configService";
 import { IFileService } from "src/code/common/service/fileService/fileService";
 import { ServiceDescriptor } from "src/code/common/service/instantiationService/descriptor";
 import { IInstantiationService, IServiceProvider } from "src/code/common/service/instantiationService/instantiation";
