+++ src/code/platform/configuration/common/configModel.ts
