+++ src/code/platform/configuration/common/configServiceBase.ts
@@ -3,7 +3,7 @@ import { Emitter, Register } from 'src/base/common/event';
 import { DataBuffer } from 'src/base/common/file/buffer';
 import { URI } from 'src/base/common/file/uri';
 import { ILogService } from 'src/base/common/logger';
-import { IConfigChange, IConfigChangeEvent, IConfigModel, IConfigType } from 'src/code/common/service/configService/configModel';
+import { IConfigChange, IConfigChangeEvent, IConfigModel, IConfigType } from 'src/code/platform/configuration/common/configModel';
 import { IFileService } from 'src/code/common/service/fileService/fileService';
 
 export interface IConfigService extends Disposable {
