+++ src/code/platform/configuration/electron/configService.ts
@@ -4,8 +4,8 @@ import { join, resolve } from "src/base/common/file/path";
 import { URI } from "src/base/common/file/uri";
 import { ILogService } from "src/base/common/logger";
 import { MarkdownRenderMode } from "src/code/browser/workbench/workspace/markdown/markdown";
-import { ConfigModel, IConfigType } from "src/code/common/service/configService/configModel";
-import { ConfigServiceBase, IConfigService } from "src/code/common/service/configService/configServiceBase";
+import { ConfigModel, IConfigType } from "src/code/platform/configuration/common/configModel";
+import { ConfigServiceBase, IConfigService } from "src/code/platform/configuration/common/configServiceBase";
 import { IFileService } from "src/code/common/service/fileService/fileService";
 import { createDecorator } from "src/code/common/service/instantiationService/decorator";
 import { IEnvironmentService, IMainEnvironmentService } from "src/code/platform/environment/common/environment";
