+++ src/code/platform/status/electron/mainStatusService.ts
@@ -3,7 +3,7 @@ import { join } from "src/base/common/file/path";
 import { URI } from "src/base/common/file/uri";
 import { ILogService } from "src/base/common/logger";
 import { mockType } from "src/base/common/util/type";
-import { NOTA_DIR_NAME } from "src/code/common/service/configService/configService";
+import { NOTA_DIR_NAME } from "src/code/platform/configuration/electron/configService";
 import { IFileService } from "src/code/common/service/fileService/fileService";
 import { createDecorator } from "src/code/common/service/instantiationService/decorator";
 import { IEnvironmentService, IMainEnvironmentService } from "src/code/platform/environment/common/environment";
